refactor: move medication to feature first folder structure
