+++ lib/core/routes.dart
@@ -10,9 +10,9 @@ import 'package:sickler/screens/emergency/crisis_logs_screen.dart';
 import 'package:sickler/screens/emergency/emergency_screen.dart';
 import 'package:sickler/screens/global_components/bottom_nav_bar.dart';
 import 'package:sickler/screens/home/home_screen.dart';
-import 'package:sickler/screens/meds/add_edit_meds_screen.dart';
+import 'package:sickler/features/meds/screens/add_edit_meds_screen.dart';
 import 'package:sickler/screens/meds/meds_details_screen.dart';
-import 'package:sickler/screens/meds/meds_schedule_screen.dart';
+import 'package:sickler/features/meds/screens/meds_schedule_screen.dart';
 import 'package:sickler/screens/meds/meds_screen.dart';
 import 'package:sickler/screens/onboarding/onboarding_base_screen.dart';
 import 'package:sickler/screens/profile/profile_basic_info_screen.dart';
