+++ lib/features/auth/screens/onboarding/components/gradient_animation.dart
